Call watchers when all of the multi operations are fired.
MockDirectory's watchers are called when each node get touched.  In
case of multi operation, this is a problem because multi is done
atomicallly, and the watcher should be fired once all of the
operations in the multi is completed.  This patch adds a flag to
propagate the caller, and if the caller is multi, evacuate the
watchers.

Signed-off-by: Yoshiaki Tamura <yoshi@midokura.com>
Acked-by: Pino de Candia <gdecandia@midokura.com>