Fixed OpenvSwitchDatabaseConnectionImpl to use getValueAsText instead of toString and check the error property of the response JSON object appropriately.
This patch fixes doJsonRpc to use getValueAsText instead of toString because
toString serializes "connection_lost" into "\"connection_lost\"". I chose
toString because response.get("id") could be string or number and I could't
know which type the `id` is until the runtime. However, getValueAsText returns
the string representation even if the `id` is number object. So I fixed it to
use getValueAsText This also fixes doJsonRpc to check whether the error
property of the response is `null` in JSON format or not. If it is `null`, the
error handling block do nothing.

Fixes #434
CR: 374, 402

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>
Acked-by: Yoshi Tamura <yoshi@midokura.com>