Fix adding a SEQUENTIAL node whose prefix ends with "/".
When String.split() is called w/ limit=0, or w/o limit specified,
trailing empty strings are not included in the resulting array.  Since
addPersistentSequential() in ZkManager postfixes a path with "/", this
causes a problem creating a node in unexpected path.  By calling
split() with limit=-1, trailing empty strings are included in the
array, and therefore fixes the problem.

Signed-off-by: Yoshiaki Tamura <yoshi@midokura.com>
Signed-off-by: Pino de Candia <gdecandia@midokura.com>