Fix broken MainTest by introducing TenantClient.java
Since baseUri requested by getBaseUri() is not set for client and leads to test failure, TenantClient
which extends Tenant and has bridges and routers variables for deserialization is introduced.
JAXB seems to need those variables to be public maybe because of broken inheritance.
TenantClient is meant to be only for testing.

CR: 410

Signed-off-by: Tomoe Sugihara <tomoe@midokura.com>
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>
Signed-off-by: Pino de Candia <gdecandia@midokura.com>