Bgp Smoke test completed.
Enhanced the API to allow the retrieval of the routes from a certain router.
Added TimedHelper to allow the execution of a job every x millis until a maximum of a y millis has passed.
Changed the JsonJaxb provider to be used in the MockMidolmanMgmt class to the same one used on the server side (this depends on a commit on the midolmanj-mgmt for now).

Signed-off-by: Toader Mihai Claudiu <mtoader@midokura.com>