Allow the midolman-mgmt application to ignore unknown properties in incoming json requests.
Added a new provider of MessageBodyWriter/Reader (a subclass of JacksonJaxbJsonProvider)
which is configured to IGNORE_UNKNONW_PROPERTIES when converting a JSON representation
to an object.
Changed the base class of the WildcardJacksonJaxbJsonProvider to be the newly added class
and also made the WildcardProvider a proper provider by adding the missing annotation.

Signed-off-by: Toader Mihai Claudiu <mtoader@midokura.com>
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>