Provide JMX interface to fetch ARP cache entries
Create a MBean interface with a method for fetching a String
representation of a port's ARP cache's entry for a given IP number.
Provide a method implementation on Router to implement that interface.
Test that the ARP cache is available through JMXConnectorServer, and
that that's updated when an ARP reply comes in.

CR: 423
Refs: #344