Fix junit dependencies.
Some of the dependencies of this package had implicit dependencies of a junit:junit jar. This jar includes and older version of hamcrest which broke the description functionality of hamcrest. The effect was that instead of seeing proper assertion failed description like:

java.lang.AssertionError: The VM should have a proper MAC address
Expected: a String matching the regular expression "(?:[0-9a-f]{2}:){5}[Z]{2}".
     but: was "52:54:00:8f:f1:5a"

   one would see just an method not found org.hamcrest.Matcher.getDescription()

 Signed-off-by: Mihai Claudiu Toader <mtoader@midokura.com>