Allow midolmanj to start even if it can't access zserv.api socket.
Midolman will start with reduced functionality (no support for BGP at all) if it can't access/create the zebra control socket. This functionality can be controlled via a setting in the midolman.conf file ("disable_bgp_on_failure" under the midolman stanza). The default value for this setting is false to keep the current functionality intact.
A NullPortService implementation (which will only log the method calls with an warn level) is used insted of a BGPPortService to allow midolman to not raise NPE's.

Signed-off-by: Mihai Claudiu Toader <mtoader@midokura.com>
Signed-off-by: Pino de Candia <gdecandia@midokura.com>