Add option to start the BGP endpoint internal service
The internal BGP daemon is started only when a switch is connected to midolmanj.
In a two (or n) midolman daemons setup the order in which the switch would connect is essentially random.
As such the only way to truly decide one of the daemons (since it can be only one given the quagga package limitations)
has the BGP functionality enabled is to force it in the configuration file.
Also if at any point the BGP service startup fails the midolman daemon will
keep going (as opposed of bailing out like it does atm).