Fix OpenVPN failing to start when OPENVPN_SERVER is set.
40c1342e6d93cee6915639be5dff19998a990ad7 added "--proto" for all
protocols.  However in case of VpnType.OPENVPN which uses UDP, the
argument after "--proto" was missing and openvpn couldn't run at all.
This problem didn't show up in the midonet-smoketest since it was only
using TCP as of midonet-smoketest:
5989315d7703ccdd986e186feaf0171641c0bcf6.

The patch fixes the problem by appending "--proto udp" when
OPENVPN_SERVER or OPENVPN_CLIENT is set.  Also adds logs the arguments
passed to openvpn, improve the test by checking whether the process
exists after start(), and a test for starting TCP server.

Signed-off-by: Yoshiaki Tamura <yoshi@midokura.com>
Acked-by: Pino de Candia <gdecandia@midokura.com>