Deleted the unnecessary warn messages for the response from OVSDB.
This patch removes the unnecessary warn messages display the error fields of
the responses from OVSDB. According to the specification of JSON-RPC, "error
field must be null if there was no error". If the error field is not null other
warn messages show up and otherwise we get no error, which means that we get
the correct response and don't need to handle error field in it.

http://json-rpc.org/wiki/specification

CR: 481

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>
Acked-by: Yoshiaki Tamura <yoshi@midokura.com>
Acked-by: Jacob Mandelson <jlm@midokura.com>