Fix NPE when looking for a deleted port in OVSDB
When you reboot a machine in nova you will get this sequence of notifications for the machine tap port:
 [ Modify, Delete, Add, Modify]
Sometimes while we are processing the Modify operation nova will actually delete the port from OVSDB so
we end up looking into OVSDB at a port that doesn't exist anymore. The code was assuming that the port
should exist so it triggered a NPE.
(http://redmine.midokura.com/redmine/issues/506 is a production occurence of this issue).
This fix will test if a port exists when looking for the external id associated with a port.
Also there are some minor cleanups in there too.

Signed-off-by: Mihai Claudiu Toader <mtoader@midokura.com>