Make sure the midolmanj-mgmt jar is also build even when running the cobertura:cobertura target.
The cobertura:cobertura target will execute the compile target to instrument the classes and when that happens the jar is deleted (mgmt is actually by default a war and the jar output is sort of a hack). This fixes it.
Also some cleanups in the smoketest pom.xml to remove hard coded dependency on zookeeper and inherit it from mgmt -> midolman.

Signed-off-by: Mihai Claudiu Toader <mtoader@midokura.com>