Improved the unit tests for PathService.  It checks for two things:
1. The root paths, broken down into sub-paths, are inserted
into the output list correctly for both midolmanj and
midolmanj-mgmt.  If midolmanj root is "/foo/bar" and
midolmanj-mgmt root is "/foo/baz", then the the root paths
should be "/foo", "/foo/bar", and "/foo/baz".

2. All the root paths for each resource are created.  This just
verifies that the correct methods were invoked so to safeguard
against accidental deletion or modifications.

Signed-off-by: Ryu Ishimoto <ryu@midokura.com>