Fix the statistics collection.
Exposed the reactor from the VRN Controller class.
Split the stats methods in the AbstractController into send/get pairs (with proper comments).
Updated the TestBridgeOVS.scala to use the new signatures and cleaned it up a little.
Made the VifMetrics to use the midolman event loop to queue the sending of the statistics requests.

 Signed-off-by: Mihai Claudiu Toader <mtoader@midokura.com> (cherry picked from commit 5e35b47)