Handle ZooKeeper session expiration
Do not create a new ZkConnection instance upon session expiration,
because the old instance is still remaining in the existing
ZkDirectory instances which could refer to the dangling session.
Instead, create a new ZooKeeper instance; this is definitely needed
because the old instance never works again once its session is expired.
Now ZkConnection is used as placeholder for ZooKeeper instances.

Signed-off-by: Takayuki Usui <takayuki@midokura.jp>