fix race in TunnelManagementTestCase
The race exists because the test expects two messages which don't
always arrive in the same order. The solution is to stop reading
one of them (LocalPortActive) from that probe and, instead, read it
from another channel: the event stream.

This opens up another race. Next, test expects two replies from the
same probe. But upon expecting those replies, the above LocalPortActive
which we stopped reading, may or may not be on the way. The solution is
the creation of a new fishForReplyOfType in MidolmanTestCase.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>