simplify the mapping reference synchronization by using atomics ops only
natRef() does not increment if the refcount is zero. Do this through a
conditional increment operation that java's AtomicInteger does not include.

This change makes all ops atomic so reasoning about this is much easier.
It was the "increment+decrement" thing that was a) complicating things
and b) broken to start with.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>