ClusterManager and subclasses handle ZK disconnects and errors
they all use a ZkConnectionAwareWatcher injected in the base class and
a set of convenience base classes for callbacks and watchers that make
use of the connection watcher.

ZkConnectionAwareWatcher gets new handleError() that ClusterManagers
and potentially any other ZK user class can use to centralize the
decisions on how to handle errors.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>