PacketContext: keep the original match and use it to build the conntrack key
On return flows, conntrack keys should be built based on the original packet,
as it entered the virtual topology, for that, we need to original match. Note
that this comes at no extra cost, the original match was already being cloned.

The reason for this is that forward flows build the conntrack key based on
the packet resulting at the end of the simulation, as it would egress the
virtual network. On the return flow, we should match on the same.

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>