Added clean up code for BGP leftovers
When midolman ends in an unexpected manner, it is possible that part of the BGP infrastructure (bgpd, network namespaces, helper interfaces) are not deleted properly. These old elements can conflict with the ones created by a new execution. For instance, you won't be able to launch a bgpd process if another one is already running and using the same sockets.

The changes have been focused mainly in RoutingHandler. Upon the creation of a new BGP instance, that part of the code will take care of deleting old stuff still hanging around. But as there were some problems when deleting some elements (network namespaces while they were being used) we don't delete everything. Instead we detect if the old elements are still there and don't try to create or delete them, just configure them again.

A functional test has been added to check this new functionality.

There was another problem were midolman shutdown wasn't ocurring in a proper manner and we wanted to address that. I've checked that when RoutingManagerActor receives an stop/shutdown, all the BGPs are properly shutdown (there's a functional test for that too). If the shutdown doesn't happen in a correct manner, we cannot fix that in the RoutingManagerActor shutdown process, but has to be cleaned upon the creation of a new BGP instance. That has been covered in this commit, as I mentioned before.

Added changes from review comments.

References: issue #210

Signed-off-by: Abel Navarro <abel@midokura.com>