Added the URI examples for the REST APIs.
The current document doesn't contain any URI examples for the REST APIs, which
forces the developers, including the CP developers, to investigate the API with
curl. This patch adds the URL examples for all REST APIs provided in the Caddo
release and provides the obvious usage of the REST APIs to the developers.

This fixes #177.
Refer #259.

Signed-off-by: Taku Fukushima <tfukushima@midokura.com>