Added a way to override the default base URI.
The base URI for th REST API can now be set by the administrator
in web.xml.  This would be useful if the location of the API server
is different from that of the actual server that the clients are
sending requests to, such as when using a proxy server.

Signed-off-by: Ryu Ishimoto <ryu@midokura.com>