Added logging for external daemons
External daemons didn't have the possibility for their output to be logged as well as captured. Bgpd has been changed to support this new logging method, but notice though that bgpd doesn't output anything one the console (stdout or stderr).

This mechanism to drain the output has an intrinsic delay since the process starts until the drain starts. It is only possible to capture the output once the process starts, so that delay seems unavoidable.

Added changes from pull request comments (#385).

Reference: issue #163

Signed-off-by: Abel Navarro <abel@midokura.com>