Made midokura-util/pom.xml to work on Linux or Mac
With a fresh code checkout, the compilation under Mac would fail. In midokura-util/pom.xml we differentiate between Mac and other OS compilation, but in the end we do the same thing.

The pom.xml has been changed to do exactly the same whether under Linux or Mac.

Reference: issue #332

Signed-off-by: Abel Navarro <abel@midokura.com>