Fragmentation: workaround for some Linux versions
Midolman's handling of ip fragments is somewhat special:

  * It returns an ICMP FRAG_NEEDED for the first fragment in a flow.
  * The sender of the ICMP FRAG_NEEDED appears to be the final
    destination of the original packet. This is counterintuitive and
    should never happen in practice.
  * If a sending host, as opposed to a router, is sending fragmented
    packets, Midolman will suggest an mtu that is not actually lower
    than the size of the first fragment.

The last of the above points triggers an undesired behaviour with
some versions of Linux. Those versions will compare the suggested mtu
with the length declared in the ip header. In a normal PMTU scenario
the latter should be bigger, not so in Midolman's case. In this
scenario, those Linux versions will try to guess a working lower
MTU.

To avoid this, this patch increases, by one, the length in the ip
header when crafting the ICMP FRAG_NEEDED error. This fools linux
hosts into taking the suggested mtu value as valid.

Linux code can be analyzed in tag:v3.0 file:net/ipv4/route.c
function:ip_rt_frag_needed().

Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>