Changed the port unlinking operation to use DELETE HTTP verb
Port unlinking is now performed by sending a DELETE request
on the port link URI.  Previously, this was done by doing a
POST on the link URI with peerId set to null, and this was
confusing and prone to accidental unlinking.  DtoLink is
created to represent the 'link' DTO on the client side.

Change-Id: I6566cd073ecbde580da6f27b95d906d57d7315ea
Signed-off-by: Ryu Ishimoto <ryu@midokura.com>