Some DatapathController logic moved to VirtualPortManager class.
This commit changes the DatapathController's internal structure by
documenting and delegating some of its logic to a non-thread-safe
VirtualPortManager that handles:
- determining when datapath ports should be added/removed
- tracking vport up/down status

Prior to this commit, a local exterior virtual port was declared
active as soon as its datapath port was created (or known to exist).
The VirtualPortManager in this commit consistently combines updates
to the:
- InterfaceDescription lists
- datapath ports
- interface-to-vport bindings

The VPM also manages the corresponding vport and datapath life-cycle.
For example, a vport is first declared locally 'active' only when all
three of the following are true:
- the host has a vport-to-interface binding for the vport
- the InterfaceDescription for the bound interface indicates that the
  it's up (both hasLink and isUp)
- the interface has been added as a port on the local datapath

Similarly, a vport that is considered locally 'active' will
transition to 'inactive' when any of the following occur:
- the host's vport-to-interface binding is removed (in which case the
  VPortManager may also request that the datapath port be removed).
- the interface's state goes to down (either
  InterfaceDescription.hasLink or isUp become false)
- the datapath port is removed (in which case the VPortManager may
  request that the port be re-added).

Change-Id: I976e69666f4871b1e848501117545e2bf38e2e54
Signed-off-by: gdecandia <gdecandia@midokura.jp>