ClusterChainManager: watch for rule updates
MM's rules cannot be changed, except that inserting a new rule in the middle
of a chain pushes the other rules back by one position. Because
ClusterChainManager was not watching for rule updates, rule insertion was
leaving already loaded chains in an inconsistent state.

This patch makes ClusterChainManager listen and react to rule updates, just
like any cluster manager, but with a slight twist. Inserting a rule at the
beginning of a chain can cause lots of updates, as many as rules are there
in the chain. This is not desirable because of the load it induces and because
those updates are partial and would create inconsistent chains until the last
update arrives.

To solve this, we take advantage of the fact that MM enforces that rule
positions are always compact, disallowing gaps between them. This means that
in a chain with 4 rules, those rules must have positions 1, 2, 3 and 4. Any
other combination is impossible. The ClusterChainManager checks for this
property in the list of positions, whenever it's not met it assumes that this
update is part of a larger batch, and the chain builder should not be called
just yet.

Fixes #472.

Change-Id: I1a9254ab4f1985160254828e2763919efbff97a7
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>