Enable L2FilteringTest and add LLDP unit test
The test was creating LLDP frames with illegal type values for mandatory
fields which messed up deserialization and broke asserts. The
FunctionalTestsHelper uses now the right values. LLDP will also verify
that the mandatory fields are present on deserialization and throw a
MalformedPacketException otherwise.

The test was also breaking because it didn't consider that ARP replies
might be broadcasted on the bridge, leaving packets lingering on taps
that would break later assertions. A new method in the
FunctionalTestsHelper allows checking the reply on the expected tap
while also cleaning the other taps if the reply appears on them.

Change-Id: Ic465ca2ae2a918bf7090a48fd432a17d281296ed
Reference: #484
Signed-off-by: Galo Navarro <galo@midokura.com>