Hack functional tests task to run as root
Functional tests need to run as root in order to be able to create netlink
sockets. This hack takes advantage of the jvm property in the maven surefire
plugin that allows to specify a different jvm executable than the one used by
maven. This way, we tell it to run the sudoJava executable, which in fact is
just a script that runs "sudo java $*".

The script also needs to chown all generated targets to the user
executing the job so that it can clean them on the following execution.

This relies obviously on sudoers being appropriately configured on the
host so that the user running the job can execute /usr/bin/java without
a password.

Change-Id: I1fd29d3810a094451e79b474bdcfc1bf2c4a793d
Reference: #484
Signed-off-by: Galo Navarro <galo@midokura.com>