Allow router delete if no ports are bound/linked.
If the router has exterior ports that are bound to a host interface the
delete fails and returns HTTP code 409 (Conflict) and a message that
includes the binding's details.

If the router has interior ports that are linked, the delete fails and
returns HTTP code 409 and a message that includes the link's details.

If these conditions are met, the router state is deleted, including
ZooKeeper directories for the router's ARP cache and SNAT lease blocks.

This fixes #488.

Change-Id: I533c112b323ba3b1f4de5c886daea278b205c33c
Signed-off-by: gdecandia <gdecandia@midokura.jp>