Fixed TunnelPortOptions endianess serialization
In the TunnelPortOptions serialization, it was explicitely introduced
the endianess for the IP fields. The serialization didn't have the same
endianess for those fields and it has been applied. It must be remarked
that the IP fields need big endianess regardless of the endianess of the
other fields.

It was noticed that Builder's serialization was wrong for byte and
short values, because the Netlink messages are 4-byte aligned, so
padding was required for that. This was not discovered before because
we do not serialize any byte or short values in other classes.

For the test, we have user a Builder's constructor that passes a
ByteBuffer as an argument. In this case, the builder's byte order was
not set, and that made the serialization incorrect. Now the builder
takes the byte order from the buffer.

A unit test has been added to check the serialization/deserialization
of TunnelPortOptions.

OvsPortsCreateAndEnumerateTest and OvsPortsGetTest have been updated
to match the expected endianess. The expected netlink messages in the
tests are hardcoded, so they needed to be updated.

Notice that these hardcoded messages were an actual netlink capture
made when the test was created. It may happen that OVS kernel module
actually receives IP addresses as big endian and returns them as little
endian. At least the functional tests related to this seem to work.

GreTunnelTest functional test has been run and checked.

Reference: issue #173

Change-Id: I3888405d03b3c2be00c9c0833a68a94fe953b660
Signed-off-by: Abel Navarro <abel@midokura.com>