FlowController: detect lost dp flows on PacketIn
If a packet is received by the FlowController and the FlowManager
already has a dp flow matching the packet in the table, there may be
2 possible causes:

 * The packet was badly installed or deleted by a program external to
   midolman.
 * The flow installation request sent to the datapath raced with the
   datapath sending a new packet up to userspace.

The second cause is legitimate and poses no problem: the packet will
be executed with the action list cached by the FlowManager and the
datapath will get around to installing the flow, preventing new
packets from making up to userspace.

However, the 1st cause produces an inconsistency between the datapath
flow table and the FlowManager's. One second after flow installation
we assume the race condition could not be the cause and reconciliation
is necessary. We reconcile both sides by re-installing the dp flow.

To support this feature the flow table in the FlowManager was extended
to include the Flow's creation timestamp.

Fixes #500

Change-Id: Ifaa1718331e14aa16615a083d416130e59b659e9
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>