Enable FilteringConnTrackingTest
Requires various fixes due to API changes, and adapting the test to
consider the possibility of ARP reply broadcasts on a bridge due to race
conditions between the simulation and ZK on the MAC-port map.

Some functionality has been unified and moved off to the
FunctionalTestsHelper and TestBase in order to reduce duplicate code
although further improvements will be done in this area later on.

There is a minor fix on the sudoJava script that is used to run the
func. tests as root within the maven task, it was assuming
jenkins-specific paths that broke the build in other environments.

Change-Id: If2720c6ab4f0f4804ebc532b0ab7c2296a39ebd3
Reference: #484
Signed-off-by: Galo Navarro <galo@midokura.com>