Enable LinksTest and GreTunnelTest
Including a refactor of the LinksTest fixing various bugs and problems
from the latest version, and hopefully simplifying the code a bit. Some
extra info is added to the TapWrapper to ease debugging.

There is a TODO at the end of the test that is mostly redundant (the
test already brings down a tap, verifies effect on routes, restores it
and checks that routes are restored) but uses a different method to
bring down the tap. Interestingly, it looks like in this case MM works
correctly but the packet is lost in the datapath and never surfaces
where expected. I tried to find the cause with Guillermo but none of us
could understand it, so, given that the test verifies most of the
functionality I'm pushing it with a TODO to revisit later.

GreTunnelTest was just disabled for no reason.

Issue: #484
Change-Id: Ie808b133446949e80b53858ae2fc148f7a9f1d15
Signed-off-by: Galo Navarro <galo@midokura.com>