Prevent test from crashing
The test is breaking on Jenkins because it assumes that the midonet
interface is DHCP configured. But it isn't, and grep commands generate
an empty output that ends up producing ArrayOutOfBoundsException.

Since the rest of the test actually works, and the exception is just
thrown when logging messages (but not on assertions), this changeset
will make the test pass. In any case we will ensure later that the
expectations about the interface are satisfied before starting the test.

Also, stop writing the dnsserver, because this seems to be breaking the
DNS on Jenkins.

Issue: #484
Change-Id: I3a748ae735e0a2c1c04bdcd6f873fb9cf5f211a4
Signed-off-by: Galo Navarro <galo@midokura.com>