Expect LocalPortActive only after ifc up
The PingTest and FloatingIpTest were expecting a LocalPortActive after
just binding the midonet interface with a port, but this is not always
true: the binding does create the association but the LocalPortActive
true is not triggered until the interface is brought up. This change
ensures that the ifup happens before the assertion.

Issue: #484
Change-Id: I67b4b73e64a0c9f232f97f824a8c146a025de3c5
Signed-off-by: Galo Navarro <galo@midokura.com>