Add unit version of LinksTest
It reproduces the same tests as the functional one. It needed to solve
a race condition between route restoring just after bringing up an
interface, and sending an ping to the network on that interface. To do
this we wanted to snoop on messages sent from the RouterManager to the
VirtualTopologyActor to confirm that the route table had been updated.
But the RM used the ".." path to the parent actor (the VTA) instead of
referencing it directly. This ended up bypassing the vtaProbe and
making us blind to the route table updates.

After discussing with @gdecandia making the RouterManager depend
directly on the VirtualTopologyActor is not a big deal at the moment.
Should this coupling need to be broken we can simply pass the parent
reference to the RouterManager in the constructor.

Change-Id: Ia8e47e89139b1a826069311c247fc76e0734618c
Signed-off-by: Galo Navarro <galo@midokura.com>