Interface sensor no longer reads dmesg
DmesgInterfaceSensor was parsing dmesg output to look for physical
devices. That worked for some physical cards with the e1000 driver, but
didn't work with other drivers.

The sensor has been susbstituded with one that looks at the sys
directory. Interfaces with a directory in /sys/devices/virtual/net can
be considered to be virtual. The endpoint will be unknown, unless the
next sensor (NetlinkInterfaceSensor) detects this is a datapath port.

Interfaces reported in /sys/class/net will be considered physical if
they weren't previously considered virtual.

Subinterfaces, like eth0:1 will still be reported unknown, because they
don't appear in the /sys directory structure.

Reference:
http://stackoverflow.com/questions/7996524/is-there-a-proper-way-to-check-for-file-directory-existence-in-java

Reference: issue #521

Change-Id: I8416d1b434906c0f54a7d060dc101fe03629ebbb
Signed-off-by: Abel Navarro <abel@midokura.com>