This is a combination of 4 commits.
1) Make VirtualTopologyManager device maps immutable.

This commit substantially reduces the number of Akka messages that are
passed between actors in order to simulate a packet moving through the
virtual topology.

This is achieved by converting VirtualTopologyActor's device maps
(used to map UUID to {Bridge, Chain, Port, Router}) to immutable maps
exposing them via a volatile variable in the VTA's companion object.
The volatile variable holds a case class with 4 arguments, one for each
of the device maps. Whenever the VTA receives a device update, it makes
a new immutable map that differs from the previous one in just that
key, value pair, and makes a new case class to set the value of the
volatile variable.

The VTA's companion object also has a new expiringAsk function, that
replaces the one in Coordinator; it first checks the volatile
variable's contents for the device that is being 'asked'. If the device
is found, a Promise.success is created, and no ask performed. If the
device is not found in the volatile's contents, as ask is performed on
the VTA. Former clients of Coordinator.expiringAsk have been converted
to VTA.expiringAsk - as have been some clients of VTA.ask.

A flow that only accesses virtual topology elements that have already
been loaded, will therefore not trigger any 'asks' to the VTA.

2) SimulationController is no longer an actor.

This commit changes SimulationController from an actor to an internal
component of DatapathController. DC delegates PacketIn and
EmitGeneratedPacket handling to SC. SC is still defined in its own
file, even though it's a very short class, just to reduce the amount
of code changing, but a future commit should entirely remove this
class.

Apart from the stylistic clean-up, removal of the SimulationController
reduces by 1 the number of Akka messages sent per simulation.

3) Handle CheckExpiredMacPorts Akka message.

BridgeManager's preStart now schedules an instance of
CheckExpiredMacPorts (instead of the type object itself) to be
delivered to the actor every few seconds. The actor's receive method
matches on the instance, not the type.

TunnelZoneManager's GreTunnelZoneBuilder and CapwapTunnelZoneBuilder
no longer send the TunnelZone instance as a message to the
VirtualToPhysicalMapper because it is not handled (or needed) there
or in the DatapathController.

4) Make BridgeConfig and RouterConfig case classes.

Change-Id: I308000065b0902f2244ab76b1a40b4a6cf46224f
Signed-off-by: gdecandia <gdecandia@midokura.jp>