This is a combination of 3 commits.
1. New test in BridgeInvalidationTest.

MockOvsDatapathConnectionImpl can now register a listener for datapath
flow create/delete operations. MidolmanTestCase registers as such a
listener and upon receiving the callbacks publishes FlowAdded and
FlowRemoved messages to the event stream. The new test in
BridgeInvalidationTest has a probe that receives these messages so
that the test can make assertions about how the FlowController is
interacting with the Datapath layer.

Signed-off-by: gdecandia <gdecandia@midokura.jp>

2. Remove RCUDeviceManager and level of indirection.

The RCUDeviceManager and its subclasses were doing nothing more than
creating builders to pass to clusterDataClient. The builders in turn
sent messages directly to the VirtualToPhysicalMapper when they are
updated. Therefore, RCUDeviceManager has been replaced by a trait
named DeviceHandler that has a handleDevice method. Its implementations
create the builders.

Signed-off-by: gdecandia <gdecandia@midokura.jp>

3. Make Coordinator a Runnable.

Since the DatapathController was creating the Coordinator and calling
its simulate method directly, this was blocking the DatapathController
thread. Now the DatapathController creates the Coordinator (now a
Runnable) and pushes it to the ExecutionContext for execution on a
separate thread.

DatapathController should be less of a bottle-neck now.

Change-Id: I2457aa4f51ce7af92034dfc04609eb10d027387a
Signed-off-by: gdecandia <gdecandia@midokura.jp>