DatapathController handle tap deletions
When a tap is deleted the corresponding datapath port gets destroyed. The
DatapathController performs the interface scanning every 2 ms. The
Netlinkinterfacesensor sets the endpoint for all the ports of the dp to
DATAPATH. If the datapathcontroller sees an interface that is bound to
a vport but whose endpoint is TUNTAP, it recreates the dp port.
Some bug fixes:
 1) If the dp realizes that a tap was deleted the binding to the vport
 is not removed
 2) In the callback of port deleted we check if the error is "no such
 device". If that's the case the deletion is considered successful and
 the internal maps are cleaned.

Signed-off-by: Rossella Sblendido <rossella@midokura.com>

Change-Id: I9a5e1959a83b260f818fd4bfde01e7489b6d3b68

Conflicts:

	midonet-functional-tests/src/test/java/com/midokura/midonet/functional_test/LinksTest.java
	midonet-functional-tests/src/test/java/org/midonet/functional_test/FunctionalTestsHelper.java