ICMP through NAT
Currently it's impossible to let ICMP packets pass through NAT to an
external address. For exampe, if two machines inside the network ping
the same external address, the outbound translation could be performed
correctly, but it would be impossible to route the replies to the
correct source.

A typical solution for this problem is matching on ICMP-specific fields,
but these are not suported by OVS: this change implements logic inside
MM to support the feature.

Implement support for ICMP in NAT rules.
----------------------------------------

This is achieved by using the ICMP identifier from the echo request and
reply as a L4 port. Both the original transport source and destination
sent to the NatLeaseManager (NLM) as "ports". When a NAT translation is
allocated for ICMP, the NLM will accept both as valid mappings without
translating them.

Implementing these changes required altering the NAT rules to inject the
ICMP identifiers as ports to the NatLeaseManager. Second, the NLM
becomes aware of the protocol involved in the translation. The logic for
TCP/UDP packets remains unchanged.

As a result, for an ICMP request from A to Z, with a router R in the
middle, we would write the following key-value pairs to Cassandra:

- forward-key: (A, id, Z, id), representing (ipSrc, tpSrc, ipDst, tpDst)
- forward-value (R, id)
- return-key: (R, id, Z, id)
- return-value: (A, id)

Note that we are not translating the identifier and thus just relying on
the source's random choice of identifier to avoid conflicts/collisions.

If both A and B send ICMP(src, dst, icmp_type, id) to Z accross the R.

    A sends ICMP(A, Z, req, x), R translates to ICMP(R, Z, req, x)
    B sends ICMP(B, Z, req, y), R translates to ICMP(R, Z, req, y)

The reply messages would be reverse-translated accordingly:

    Z sends ICMP(Z, R, reply, x), R translates to ICMP(Z, A, reply, x)
    Z sends ICMP(Z, R, reply, y), R translates to ICMP(Z, B, reply, y)

Support for ICMP errors is implemented by examining the contents of the
ICMP error data field to extract the correct tpSrc and tpDst that can be
used to perform the reverse mapping and replace the appropriate fields
in the packet to make it reach the correct destination. The contents of
the data field can be a TCP/UDP packet (from which we'd extract the
ports), or an ICMP packet (from which we'd extract the identifier). From
this point the reverse mapping can be performed as described above.

It's worth noting that performing NAT on ICMP error messages not only
involves translating the src/dst IP, but also the src/dst 4-tuples
inside the ICMP data field.

Support for non-OVS compatible FlowKey
---------------------------------------

There are two new FlowKeys created to support the functionality
described.

FlowKeyICMPEcho allows matching and altering the ICMP identifer for both
ECHO_REQUEST and ECHO_REPLY ICMP packets. FlowKeyICMPError allows
matching on UNREACHABLE, PARAMETER_PROBLEM, TIME_EXCEEDED errors.

Any Flow that contains such key in its FlowMatch object cannot be
installed in the kernel since OVS does not offer support for those
extended fields, but should generate a WildcardFlows allowing practice
to avoid performing a full simulation.

It's reasonable to assume that the same model could be applied to other
type of traffic, where MM provides extended matching capabilities on top
of OVS. Hence, a new interface FlowKey.UserSpaceOnly can mark FlowKeys
that are intended only for userspace matches. The FlowController will
not install any flows whose match contains any FlowKey.UserSpaceOnly
key.

Issue: #513, #435
Change-Id: I5dab6aa3de02dc7a90a484175846ef63b4d1a6cb
Signed-off-by: Galo Navarro <galo@midokura.com>