Netlink: move writes to select loop
Motivation: the netlink channel is set to non-blocking because
reads are performed out of the select loop. Writes were not
checking for writeability of the channel, now this check
is done by the select loop.

Implementation: A new BlockingQueue subclass is used to queue the
requests that need to be written, this class (SelectorInputQueue)
wakes up a Selector when an item is added.

The select loop can now take multiple registrations for the same
channel (for different ops obviously), and some of them may be
bound to a BlockingQueue. For these new registrations, the loop
is interested in the ops only when there's data in the queue.

As a consequence the select loop now reconstructs the interest
set before each call to select(), and demultiplexes the ready
ops on return to find the callback associated with each of them.

Like reads, writes are done in batches, the callback writes
as many msgs as possible until either the queue is depleted or
the channel will not take any more. However, we introduce a
maximum number of operations per batch to prevent one operation
from adding too much latency to the other.

Change-Id: I595b736fc7fd3b26e1007da0b722fcf59913579a
Signed-off-by: Guillermo Ontanon <guillermo@midokura.jp>